Move into massa-versioning-worker && massa-versioning-exports
