+++ Cargo.toml
@@ -36,6 +36,8 @@ members = [
   "massa-final-state",
   "massa-pos-exports",
   "massa-pos-worker",
+  "massa-versioning-worker",
+  "massa-versioning-exports",
 ]
 resolver = "2"
