+++ massa-models/Cargo.toml
@@ -18,8 +18,6 @@ config = "0.13"
 bs58 = { version = "0.4", features = ["check"] }
 bitvec = { version = "1.0", features = ["serde"] }
 nom = "7.1"
-machine = "0.3"
-parking_lot = "0.12"
 
 # custom modules
 massa_hash = { path = "../massa-hash" }
