+++ massa-models/src/test_exports/mod.rs
@@ -1,4 +1,3 @@
 mod tools;
-pub mod versioning_helpers;
 
 pub use tools::*;
