+++ massa-versioning-exports/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "massa_versioning_exports"
+version = "0.1.0"
+authors = ["Massa Labs <info@massa.net>"]
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[dev-dependencies]
+
+[features]
