+++ massa-versioning-worker/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod versioning;
+pub mod versioning_factory;
+
+/// Test utils
+#[cfg(feature = "testing")]
+pub mod test_helpers;
