+++ massa-versioning-worker/src/test_helpers/mod.rs
@@ -1,0 +1,1 @@
+pub mod versioning_helpers;
